fix(admin): refactor selection of relation in RelationsView
- move initial setting back out into RelationsView
-- no reason for this to happen in each RelationBox
- always reset relation when available relations change

Refs: TOCDEV-4572
