+++ packages/admin/src/routes/entities/components/RelationsView/RelationBox.js
@@ -1,53 +1,30 @@
-import React, {useEffect} from 'react'
+import React from 'react'
 import {AdminLink as StyledLink, Icon} from 'tocco-ui'
-import queryString from 'query-string'
-import _get from 'lodash/get'
 import PropTypes from 'prop-types'
 
-import {getRelation, setRelation} from '../../utils/relationPersistor'
+import {setRelation} from '../../utils/relationPersistor'
 import {StyledRelationBox, StyledRelationLabel, StyledRelationLinks} from './StyledComponents'
-import {currentViewPropType} from '../../utils/propTypes'
 
 const RelationBox = ({
   relation,
   history,
   match,
-  currentViewInfo,
-  relations,
   relationsInfo,
   selectRelation,
   selectedRelation,
-  intl
+  intl,
+  entityName
 }) => {
-  useEffect(
-    () => {
-      if (!selectedRelation && relations?.length > 0) {
-        const queryRelation = queryString.parse(history.location.search).relation
-        if (queryRelation) {
-          selectRelation(relations.find(r => r.relationName === queryRelation))
-        } else {
-          const persistedSelectedRelation = getRelation(entityName)
-          if (persistedSelectedRelation) {
-            selectRelation(relations.find(r => r.relationName === persistedSelectedRelation))
-          } else {
-            selectRelation(relations[0])
-          }
-        }
-      }
-    },
-    [relations, selectedRelation]
-  )
-
-  const msg = id => intl.formatMessage({id})
-  const entityName = _get(currentViewInfo, 'model.name')
-  const hasCreateRights = relationName => relationsInfo[relationName]?.createPermission
   const {
     relationName,
     targetEntity,
     relationDisplay
   } = relation
 
-  const getRelationCountLabel = relationName => relationsInfo[relationName]?.count > 0
+  const msg = id => intl.formatMessage({id})
+  const hasCreateRights = () => relationsInfo[relationName]?.createPermission
+
+  const getRelationCountLabel = () => relationsInfo[relationName]?.count > 0
     ? <StyledRelationLabel>&nbsp;({relationsInfo[relationName].count})</StyledRelationLabel>
     : null
 
@@ -65,14 +42,14 @@ const RelationBox = ({
       onClick={handleBoxClick}
     >
       <StyledRelationLabel title={relationDisplay.label}>
-        {relationDisplay.label}</StyledRelationLabel>{getRelationCountLabel(relationName)}
+        {relationDisplay.label}</StyledRelationLabel>{getRelationCountLabel()}
       <StyledRelationLinks>
         <StyledLink
           aria-label={msg('client.admin.entities.relationsView.relationLinkView')}
           to={match.url.replace(/(relations|detail)$/, relationName)}>
           <Icon icon="arrow-right"/>
         </StyledLink>
-        {hasCreateRights(relationName) && targetEntity !== 'Resource'
+        {hasCreateRights() && targetEntity !== 'Resource'
         && <StyledLink
           aria-label={msg('client.admin.entities.relationsView.relationLinkCreate')}
           to={match.url.replace(/(relations|detail)$/, relationName) + '/create'}>
@@ -88,15 +65,14 @@ RelationBox.propTypes = {
   relation: PropTypes.object.isRequired,
   history: PropTypes.object.isRequired,
   match: PropTypes.object.isRequired,
-  currentViewInfo: currentViewPropType,
-  relations: PropTypes.array,
   relationsInfo: PropTypes.objectOf(PropTypes.shape({
     count: PropTypes.number,
     createPermission: PropTypes.bool
   })),
   intl: PropTypes.object.isRequired,
   selectRelation: PropTypes.func.isRequired,
-  selectedRelation: PropTypes.object
+  selectedRelation: PropTypes.object,
+  entityName: PropTypes.string
 }
 
 export default RelationBox
